+++ src/DbConnectionPlus.DatabaseAdapters.Oracle/GlobalUsings.cs
@@ -1,10 +1,6 @@
 global using System.Collections;
 global using System.Collections.Concurrent;
-global using System.ComponentModel.DataAnnotations;
-global using System.ComponentModel.DataAnnotations.Schema;
 global using System.Data;
 global using System.Data.Common;
-global using System.Globalization;
-global using System.Runtime.CompilerServices;
 global using RentADeveloper.DbConnectionPlus.Configuration;
 global using RentADeveloper.DbConnectionPlus.DatabaseAdapters;