+++ src/DbConnectionPlus.DatabaseAdapters.PostgreSql/AssemblyAttributes.cs
@@ -1,0 +1,3 @@
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("RentADeveloper.DbConnectionPlus.UnitTests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("RentADeveloper.DbConnectionPlus.IntegrationTests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("DynamicProxyGenAssembly2")]