+++ src/DbConnectionPlus.DatabaseAdapters.SqlServer/GlobalUsings.cs
@@ -1,10 +1,7 @@
 global using System.Collections;
 global using System.Collections.Concurrent;
-global using System.ComponentModel.DataAnnotations;
-global using System.ComponentModel.DataAnnotations.Schema;
 global using System.Data;
 global using System.Data.Common;
-global using System.Globalization;
-global using System.Runtime.CompilerServices;
+global using Microsoft.Data.SqlClient;
 global using RentADeveloper.DbConnectionPlus.Configuration;
 global using RentADeveloper.DbConnectionPlus.DatabaseAdapters;